Organize some of the LTN business logic code
