+++ apps/ltn/src/logic/impact.rs
