+++ apps/ltn/src/logic/mod.rs
@@ -1,1 +1,13 @@
+mod auto_filters;
 pub mod crossings;
+mod existing_filters;
+pub mod impact;
+mod partition;
+mod shortcuts;
+
+pub use auto_filters::AutoFilterHeuristic;
+pub use crossings::populate_existing_crossings;
+pub use existing_filters::transform_existing_filters;
+pub use impact::Impact;
+pub use partition::{BlockID, CustomBoundary, NeighbourhoodID, Partitioning};
+pub use shortcuts::Shortcuts;
