+++ apps/ltn/src/logic/partition.rs
